Improve course-specific scrapes, fix potential 'View All' sections bug
- Edit scrape_specific_course method to take a subject & course code.
This is better than the previous where you just gave a letter and
index of the course on that page.
- Fix potential bug with 'View All' sections.  If you 'View All' for one
term's sections, and then switch terms, you are still in view all mode.
Clicking the button again would collapse the sections.  This commit
adds a check for the text content of the button to determine if you
would be collapsing or expanding